refactored token logic
