+++ .TODO
@@ -12,6 +12,5 @@
 * Сделать валидацию форм
 * Сделать для всех форм равный отступ сверху
 * Проверить оформляется ли заказ с протухшим токеном
-* Типизировать requestCreator
 * После оформления заказа и перезагрузки страницы модалка заказа открыта. Исправить
 * fetch('https://norma.nomoreparties.space/api/orders')
